SAM-1094
git-svn-id: https://source.sakaiproject.org/svn/sam/branches/samigo-2.8.x@87883 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo-app/src/java/org/sakaiproject/tool/assessment/ui/bean/qti/XMLImportBean.java
@@ -138,6 +138,7 @@ public class XMLImportBean implements Serializable
   {
 	String filename = uploadFile;
 	String unzipLocation = null;
+	boolean fileNotFound = false;
 	if (isCP) {
 		ImportService importService = new ImportService();
 		unzipLocation = importService.unzipImportFile(uploadFile);
@@ -149,6 +150,7 @@ public class XMLImportBean implements Serializable
     }
     catch (FileNotFoundException fnfex)
     {
+      fileNotFound = true;
       ResourceLoader rb = new ResourceLoader("org.sakaiproject.tool.assessment.bundle.AuthorImportExport");
       FacesMessage message = new FacesMessage( rb.getString("import_qti_not_found") );
       FacesContext.getCurrentInstance().addMessage(null, message);
@@ -160,12 +162,16 @@ public class XMLImportBean implements Serializable
       FacesContext.getCurrentInstance().addMessage(null, message);
     }
     finally {
+      boolean success = false;    	
       // remove unsuccessful file
-      log.debug("****Clean up file: "+filename);
-      File f1 = new File(filename);
-      boolean success = f1.delete();
-      if (!success)
-	log.error ("Failed to delete file " + filename);
+      if (!fileNotFound) {
+        log.debug("****Clean up file: "+filename);
+        File f1 = new File(filename);
+        success = f1.delete();
+        if (!success) {
+    	  log.error ("Failed to delete file " + filename);
+        }
+      }
       if (isCP) {
     	  File f2 = new File(uploadFile);
     	  success = f2.delete();
